requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.26.2
Jinja2==3.1.2
factory-boy==3.2.1
Faker==18.9.0
greenlet==2.0.2
marshmallow==3.19.0
packaging==23.1
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
setuptools==67.8.0
six==1.16.0
social-auth-app-django==5.2.0
soupsieve==2.4.1
sqlparse==0.4.4
SQLAlchemy==1.4.31
